update: Zennデータ取得時の整理
